+++ game_patch/multi/server.h
@@ -13,3 +13,4 @@ bool server_is_saving_enabled();
 void server_reliable_socket_ready(rf::Player* player);
 bool server_weapon_items_give_full_ammo();
 const char* get_rand_level_filename();
+void shuffle_level_array();